@EgorBo 2023-07-19
> The removal of bound checks has no visible effect on the perf (+-1%, within the range of error),

Maybe then revert that part? There is an on-going effort to reduce number of unsafe hacks (e.g. bound checks) in sake of better safety.